// src/schemas/auth.ts
export const addUserSchema = z.object({
  name: z.string().min(1, "Name is required"),
  email: z.string().regex(
    /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
    "Please provide a valid email address"
  ),
  userType: z.enum([
    "Doctor",
    "Receptionist",
    "Nurse",
    "Pharmacist",
    "Lab_Assistant",
    "Radiologist",
  ]),
  mobile: z.string().optional(),
  speciality: z.string().nullable().optional(),
  registrationNumber: z.string().optional(),
  labId: z.string().optional(),
  pharmacyId: z.string().optional(),
});
// ...
export type AddUserDto = z.infer<typeof addUserSchema>;